# docs/implementation/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/lessons": {
      "get": {
        "operationId": "PrivateController_lessons",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      },
      "post": {
        "operationId": "PrivateController_createLesson",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
    "/api/v1/lessons/{id}": {
      "get": {
        "operationId": "PrivateController_lesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      },
      "patch": {
        "operationId": "PrivateController_updateLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
...
    "/api/v1/classes": {
      "get": {
        "operationId": "ClassroomController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      },
      "post": {
        "operationId": "ClassroomController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/invites": {
      "post": {
        "operationId": "ClassroomController_invite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/join": {
      "post": {
        "operationId": "ClassroomController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}": {
      "get": {
        "operationId": "ClassroomController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/assignments": {
      "post": {
        "operationId": "ClassroomController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/assignments/{assignmentId}": {
      "get": {
        "operationId": "ClassroomController_assignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/assignments/{assignmentId}/submissions": {
      "post": {
        "operationId": "ClassroomController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/results/{assignmentId}": {
      "get": {
        "operationId": "ClassroomController_results",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/members/{userId}": {
      "delete": {
        "operationId": "ClassroomController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    },
    "/api/v1/classes/{id}/members": {
      "get": {
        "operationId": "ClassroomController_members",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classroom"
        ]
      }
    }
  },
  "info": {
    "title": "HumanScope REST API",
    "description": "Foundation route inventory. Body validation is defined by Zod contracts; response schemas and full OpenAPI coverage remain a release gate.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hs_session"
      }
    },
...
  }
}